Update Gradle, AGP, and other dependencies
Fix deprecations related to the upgrades

+++ readium/navigators/web/internals/src/main/kotlin/org/readium/navigator/web/internals/webview/WebViewScrollController.kt
@@ -169,7 +169,7 @@ private fun RelaxedWebView.scrollToProgression(
 ) {
     when (orientation) {
         Orientation.Vertical -> {
-            scrollTo(scrollX, progression.roundToInt() * maxScrollY.toInt())
+            scrollTo(scrollX, progression.roundToInt() * maxScrollY)
         }
         Orientation.Horizontal -> when (direction) {
             LayoutDirection.Ltr -> {
